requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2.12
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-six==1.0.4
djangorestframework==3.13.1
idna==3.3
pytz==2021.3
requests==2.27.1
screen==1.0.1
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.8
xlwt==1.3.0
